GDPR-111 Refactoring - separating out services into clearer responsibilities
