+++ docker-compose.yml
@@ -29,7 +29,7 @@ services:
       - AWS_SECRET_ACCESS_KEY=arbitrary
 
   oauth:
-    image: mojdigitalstudio/nomis-oauth2-server:latest
+    image: quay.io/hmpps/nomis-oauth2-server:latest
     networks:
       - hmpps
     container_name: oauth
@@ -42,7 +42,7 @@ services:
       - SPRING_PROFILES_ACTIVE=dev
 
   elite2-api:
-    image: mojdigitalstudio/elite2-api:latest
+    image: quay.io/hmpps/elite2-api:latest
     container_name: elite2-api
     depends_on:
       - localstack
