+++ src/main/java/uk/gov/justice/hmpps/datacompliance/repository/jpa/model/OffenderEntity.java
@@ -1,0 +1,10 @@
+package uk.gov.justice.hmpps.datacompliance.repository.jpa.model;
+
+import uk.gov.justice.hmpps.datacompliance.dto.OffenderNumber;
+
+/**
+ * A database entity that is linked to an offender.
+ */
+public interface OffenderEntity {
+    OffenderNumber getOffenderNumber();
+}
