+++ src/main/java/uk/gov/justice/hmpps/datacompliance/services/referral/ReferralService.java
@@ -1,0 +1,75 @@
+package uk.gov.justice.hmpps.datacompliance.services.referral;
+
+import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import uk.gov.justice.hmpps.datacompliance.dto.OffenderNumber;
+import uk.gov.justice.hmpps.datacompliance.events.listeners.dto.OffenderPendingDeletionEvent;
+import uk.gov.justice.hmpps.datacompliance.events.listeners.dto.OffenderPendingDeletionReferralCompleteEvent;
+import uk.gov.justice.hmpps.datacompliance.repository.jpa.model.referral.OffenderDeletionReferral;
+import uk.gov.justice.hmpps.datacompliance.repository.jpa.model.referral.ReferredOffenderBooking;
+import uk.gov.justice.hmpps.datacompliance.repository.jpa.repository.referral.OffenderDeletionBatchRepository;
+import uk.gov.justice.hmpps.datacompliance.services.retention.RetentionService;
+import uk.gov.justice.hmpps.datacompliance.utils.TimeSource;
+
+import static uk.gov.justice.hmpps.datacompliance.utils.Exceptions.illegalState;
+
+@Slf4j
+@Service
+@Transactional
+@AllArgsConstructor
+public class ReferralService {
+
+    private final TimeSource timeSource;
+    private final OffenderDeletionBatchRepository batchRepository;
+    private final RetentionService retentionService;
+    private final ReferralResolutionService referralResolutionService;
+
+    public void handlePendingDeletionReferral(final OffenderPendingDeletionEvent event) {
+
+        final var referral = createReferral(event);
+
+        final var retentionChecks = retentionService.conductRetentionChecks(
+                new OffenderNumber(event.getOffenderIdDisplay()));
+
+        referralResolutionService.processReferral(referral, retentionChecks);
+    }
+
+    public void handleReferralComplete(final OffenderPendingDeletionReferralCompleteEvent event) {
+
+        log.info("All offenders pending deletion in batch: '{}' have been added to the queue", event.getBatchId());
+
+        final var batch = batchRepository.findById(event.getBatchId())
+                .orElseThrow(illegalState("Cannot find batch with id: '%s'", event.getBatchId()));
+
+        batch.setReferralCompletionDateTime(timeSource.nowAsLocalDateTime());
+
+        batchRepository.save(batch);
+    }
+
+    private OffenderDeletionReferral createReferral(final OffenderPendingDeletionEvent event) {
+
+        final var batch = batchRepository.findById(event.getBatchId())
+                .orElseThrow(illegalState("Cannot find deletion batch with id: '%s'", event.getBatchId()));
+
+        final var referral = OffenderDeletionReferral.builder()
+                .offenderDeletionBatch(batch)
+                .receivedDateTime(timeSource.nowAsLocalDateTime())
+                .offenderNo(event.getOffenderIdDisplay())
+                .firstName(event.getFirstName())
+                .middleName(event.getMiddleName())
+                .lastName(event.getLastName())
+                .birthDate(event.getBirthDate())
+                .build();
+
+        event.getOffenders().forEach(offender ->
+                offender.getOffenderBookings().forEach(booking ->
+                        referral.addReferredOffenderBooking(ReferredOffenderBooking.builder()
+                                .offenderId(offender.getOffenderId())
+                                .offenderBookId(booking.getOffenderBookId())
+                                .build())));
+
+        return referral;
+    }
+}
